requirements.txt
sanic==21.9.3
aiohttp==3.8.1
PyNaCl==1.4.0